Replace generate-kcp-api.sh with Makefile targets
Instead of a fairly complicated generate-kcp-api.sh shell script, we can
now use the semi-complicated make targets. The advantage of this
approach is that we can now mandate that the `all` target includes
generating the kcp manifests.
Also the kubectl kcp plugin is now resolved via go modules (using
`tools.go`) which makes it easier to upgrade and run (using `go run`).
The GitHub Actions "Checks" workflow now makes sure that the generated
files are up to date, so we don't forget to update them.
Also the generated files are updated with the current definitions.