Simplify policy source schema
Let's just use a list of strings, each one being a go-getter style
url compatible with conftest pull.

This is related to some refactoring I'm attempting to do in the
ec-cli repo.

I think this could break clusters and e2e tests, so we need to plan
how to merge it with the least chaos.