Add IgnoreRekor attribute to ECP
https://issues.redhat.com/browse/HACBS-2165

Let's be explicit about when to skip Rekor during validation. The
RekorUrl parameter is not always required to enforce Rekor verification.
For example, if a signature contains a bundle with
SignedEntryTimestamp[0], then the RekorURL is skipped in favor of the
Rekor public keys loaded via TUF. For those use cases, leaving the
RekorUrl empty, but still expect Rekor to be part of the verification,
is in fact the right thing to do.

[0] https://github.com/sigstore/cosign/blob/main/specs/SIGNATURE_SPEC.md

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>