EC-80 Document formal spec for EC policy.yaml
This commit adds functionality to make the EC policy.yaml format self
documenting. Any changes to the EC policy structure will be reflected in
a update to the exported schema thanks to the github workflow. The
updated schema is pushed to the enterprisecontract.dev
[website](https://enterprisecontract.dev/enterprise-contract-controller/schema/policy_spec.json)
This location will be registered with schemastore.org.

Signed-off-by: Rob Nester <rob.nester@gmail.com>