Add "reference" field to volatileConfig
In the volatileConfig list item (part of the ECP), we add another field
called "reference". It is optional and does nothing, but it is a place
to add a Jira link to the issue tracker for the reason why the exception
is needed.

This API change will not affect current clients since the new field is not
required.

Ref: https://issues.redhat.com/browse/EC-1246